creating and editing posts, some home, created mapping folder
