+++ client/src/components/Mapping/ThreadMapping.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { Link } from 'react-router-dom'
 import { getThreads } from '../../services/thread'
 
-export default function ThreadSingle(props) {
+export default function ThreadMapping(props) {
   const [threads, setThreads] = useState([])
 
   useEffect(() => {
