+++ client/src/screens/Posts/createpost.jsx
@@ -49,7 +49,7 @@ console.log(props.user)
         <label>Image:</label>
         <input id="imgUrl" value={input.imgUrl} onChange={handleChange} />
         <br />
-        {/* <button ref={"refresh"}>Submit</button> */}
+        {/* <button ref={refresh}>Submit</button> */}
         <button>Submit</button>
       </form>
     </>
